reduce number of code files, fix page size in paged slice
