+++ ecs/archetype_test.go
@@ -3,7 +3,6 @@ package ecs
 import (
 	"reflect"
 	"testing"
-	"unsafe"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -240,28 +239,6 @@ func TestArchetypeZero(t *testing.T) {
 	arch.Alloc(newEntity(1))
 	assert.Equal(t, position{0, 0}, *(*position)(arch.Get(0, 0)))
 	assert.Equal(t, position{0, 0}, *(*position)(arch.Get(1, 0)))
-}
-
-func TestArchetypePointers(t *testing.T) {
-	pt := archetypePointers{}
-
-	a1 := archetype{}
-	a2 := archetype{}
-	a3 := archetype{}
-
-	pt.Add(&a1)
-	pt.Add(&a2)
-	pt.Add(&a3)
-
-	assert.Equal(t, 3, pt.Len())
-
-	for i := 0; i < 45; i++ {
-		pt.Add(&a3)
-	}
-
-	assert.Equal(t, unsafe.Pointer(&a1), unsafe.Pointer(pt.Get(0)))
-	assert.Equal(t, unsafe.Pointer(&a2), unsafe.Pointer(pt.Get(1)))
-	assert.Equal(t, unsafe.Pointer(&a3), unsafe.Pointer(pt.Get(2)))
 }
 
 func BenchmarkIterArchetype_1000(b *testing.B) {
