+++ ecs/pool_test.go
@@ -98,3 +98,37 @@ func TestEntityPoolStochastic(t *testing.T) {
 		}
 	}
 }
+
+func TestBitPool(t *testing.T) {
+	p := bitPool{}
+
+	for i := 0; i < MaskTotalBits; i++ {
+		assert.Equal(t, i, int(p.Get()))
+	}
+
+	assert.Panics(t, func() { p.Get() })
+
+	for i := 0; i < 10; i++ {
+		p.Recycle(uint8(i))
+	}
+	for i := 9; i >= 0; i-- {
+		assert.Equal(t, i, int(p.Get()))
+	}
+
+	assert.Panics(t, func() { p.Get() })
+
+	p.Reset()
+
+	for i := 0; i < MaskTotalBits; i++ {
+		assert.Equal(t, i, int(p.Get()))
+	}
+
+	assert.Panics(t, func() { p.Get() })
+
+	for i := 0; i < 10; i++ {
+		p.Recycle(uint8(i))
+	}
+	for i := 9; i >= 0; i-- {
+		assert.Equal(t, i, int(p.Get()))
+	}
+}
