+++ ecs/util_test.go
@@ -2,6 +2,7 @@ package ecs
 
 import (
 	"testing"
+	"unsafe"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -40,4 +41,32 @@ func TestLockMask(t *testing.T) {
 
 	locks.Unlock(l2)
 	assert.False(t, locks.IsLocked())
+}
+
+func TestPagedSlice(t *testing.T) {
+	a := newPagedSlice[int](32)
+
+	for i := 0; i < 66; i++ {
+		a.Add(i)
+		assert.Equal(t, i, *a.Get(i))
+		assert.Equal(t, i+1, a.Len())
+	}
+}
+
+func TestPagedSlicePointerPersistence(t *testing.T) {
+	a := newPagedSlice[int](32)
+
+	a.Add(0)
+	p1 := a.Get(0)
+
+	for i := 1; i < 66; i++ {
+		a.Add(i)
+		assert.Equal(t, i, *a.Get(i))
+		assert.Equal(t, i+1, a.Len())
+	}
+
+	p2 := a.Get(0)
+	assert.Equal(t, unsafe.Pointer(p1), unsafe.Pointer(p2))
+	*p1 = 100
+	assert.Equal(t, 100, *p2)
 }
